refactor(desktop): move daily review action helpers
