+++ apps/desktop/src/renderer/daily-review-actions.ts
@@ -1,0 +1,33 @@
+import type { LlmConnection } from '@maka/core';
+import { generalizedErrorMessageChinese } from '@maka/core';
+import { buildCatalogDailyReviewModelOptions } from './model-catalog-choices';
+
+export const DAILY_REVIEW_CONFIG_MODEL_VALUE = '__maka_daily_review_config_model__';
+
+export function dailyReviewExportDefaultName(label: string): string {
+  const today = new Date();
+  const yyyy = today.getFullYear();
+  const mm = String(today.getMonth() + 1).padStart(2, '0');
+  const dd = String(today.getDate()).padStart(2, '0');
+  const scope = label.includes('30')
+    ? '30d'
+    : label.includes('7')
+      ? '7d'
+      : label === '昨天'
+        ? 'yesterday'
+        : label === '今天'
+          ? 'today'
+          : 'day';
+  return `maka-daily-review-${scope}-${yyyy}-${mm}-${dd}.md`;
+}
+
+export function dailyReviewActionErrorMessage(error: unknown, fallback: string): string {
+  return generalizedErrorMessageChinese(error, fallback);
+}
+
+export function buildDailyReviewRunModelOptions(connections: readonly LlmConnection[]): Array<readonly [string, string]> {
+  return [
+    [DAILY_REVIEW_CONFIG_MODEL_VALUE, '使用设置中的分析模型'],
+    ...buildCatalogDailyReviewModelOptions(connections, ''),
+  ];
+}
